move columns, columnRules, editRiskAssessmentColumns as component props
